// mi-primera-codeapp/src/generated/models/Cra58_tareasModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cra58_tareascra58_estado = {
  962000000: 'Pendiente',
  962000001: 'Enprogreso',
  962000002: 'Completada'
} as const;
export type Cra58_tareascra58_estado = keyof typeof Cra58_tareascra58_estado;
export const Cra58_tareascra58_prioridad = {
  962000000: 'Alta',
  962000001: 'Media',
  962000002: 'Baja'
} as const;
export type Cra58_tareascra58_prioridad = keyof typeof Cra58_tareascra58_prioridad;
export const Cra58_tareasstatecode = {
  0: 'Activo',
  1: 'Inactivo'
} as const;
export type Cra58_tareasstatecode = keyof typeof Cra58_tareasstatecode;
export const Cra58_tareasstatuscode = {
  1: 'Activo',
  2: 'Inactivo'
} as const;
export type Cra58_tareasstatuscode = keyof typeof Cra58_tareasstatuscode;

export interface Cra58_tareasBase {
  cra58_descripcion?: string;
  cra58_estado?: Cra58_tareascra58_estado;
  cra58_fechalimite?: string;
  cra58_prioridad?: Cra58_tareascra58_prioridad;
  "cra58_ProyectoID@odata.bind"?: string;
  cra58_tareaid: string;
  cra58_titulo?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cra58_tareasstatecode;
  statuscode?: Cra58_tareasstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cra58_tareas extends Cra58_tareasBase {
  cra58_estadoname?: string;
  cra58_prioridadname?: string;
  cra58_proyectoidname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  cra58_proyectoid?: object;
  _cra58_proyectoid_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}