// mi-primera-codeapp/src/generated/models/LISTADODEWIKISModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface TecnologiaUtilizadaValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface OrigenDeDatosValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface GeneradaPorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface TipoWikiValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface EstadoValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface OpcionesValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface AuthorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface EditorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface LISTADODEWIKISBase {
  ID?: number;
  Title: string;
  IDPersonalizado?: number;
  "TecnologiaUtilizada#Id"?: number[];
  "TecnologiaUtilizada@odata.type"?: string;
  "OrigenDeDatos#Id"?: number[];
  "OrigenDeDatos@odata.type"?: string;
  "GeneradaPor#Claims"?: string;
  EsPaginaWiki?: boolean;
  CreadorWikiApp?: string;
  "TipoWiki#Id"?: number;
  Acceder?: string;
  Favorito?: boolean;
  Eliminado?: boolean;
  "Estado#Id"?: number;
  "Opciones#Id"?: number;
  OData__ColorTag?: string;
  ComplianceAssetId?: string;
  Modified?: string;
  Created?: string;
  "Author#Claims"?: string;
  "Editor#Claims"?: string;
  "{Identifier}"?: string;
  "{IsFolder}"?: boolean;
  "{Link}"?: string;
  "{Name}"?: string;
  "{FilenameWithExtension}"?: string;
  "{Path}"?: string;
  "{FullPath}"?: string;
  "{ModerationStatus}"?: string;
  "{ModerationComment}"?: string;
  "{ContentType}#Id"?: string;
  "{ContentType}"?: {
    "@odata.type"?: string;
    Id?: string;
    Name?: string;
  };
  "{HasAttachments}"?: boolean;
  "{Attachments}@odata.type"?: string;
  "{Attachments}"?: {
   "@odata.type"?: string;
   Id?: string;
   AbsoluteUri?: string;
   DisplayName?: string;
 }[];
  "{VersionNumber}"?: string;
  "{TriggerWindowStartToken}"?: string;
  "{TriggerWindowEndToken}"?: string;
}

export interface LISTADODEWIKISRead extends LISTADODEWIKISBase {
  TecnologiaUtilizada?: TecnologiaUtilizadaValue;
  OrigenDeDatos?: OrigenDeDatosValue;
  GeneradaPor: GeneradaPorValue;
  TipoWiki: TipoWikiValue;
  Estado?: EstadoValue;
  Opciones?: OpcionesValue;
  Author?: AuthorValue;
  Editor?: EditorValue;
}

export interface LISTADODEWIKISWrite extends LISTADODEWIKISBase {
  TecnologiaUtilizada?: TecnologiaUtilizadaValue;
  OrigenDeDatos?: OrigenDeDatosValue;
  GeneradaPor: GeneradaPorValue;
  TipoWiki: string;
  Estado?: string;
  Opciones?: string;
  Author?: AuthorValue;
  Editor?: EditorValue;
}